projects are created implicit and managed by UI
